multiple account login

+++ utils/fetchUserDetail.js
@@ -1,16 +1,20 @@
 export const fetchUserAccessToken = () => {
-    const accessToken = localStorage.getItem("accessToken") !== undefined
-        ? JSON.parse(localStorage.getItem("accessToken"))
-        : localStorage.clear();
+    if (typeof window !== 'undefined') {
+        const accessToken = localStorage.getItem("accessToken") !== undefined
+            ? JSON.parse(localStorage.getItem("accessToken"))
+            : localStorage.clear();
 
-    return accessToken;
+        return accessToken;
+    }
 }
 
 
 export const fetchUserInfo = () => {
-    const userInfo = localStorage.getItem("user") !== undefined
-        ? JSON.parse(localStorage.getItem("user"))
-        : localStorage.clear();
+    if (typeof window !== 'undefined') {
+        const userInfo = localStorage.getItem("user") !== undefined
+            ? JSON.parse(localStorage.getItem("user"))
+            : localStorage.clear();
 
-    return userInfo;
+        return userInfo;
+    }
 }
